requirements.txt
requests>=2.31.0
anthropic>=0.25.0
python-dotenv>=1.0.0
schedule>=1.2.0
